Continued compartmentalising scripts, started work on a unifying test script in test folder.
