+++ canarybot/test_canary.py
@@ -1,0 +1,4 @@
+import unittest
+from unittest.mock import MagicMock,patch,mock_open
+import canary
+
